Break up rateCalculations into separate files
